write_xlsx {writexl}	R Documentation

Export to xlsx

Description

Writes a data frame to an xlsx file. To create an xlsx with (multiple) named sheets, simply set x to a named list of data frames.

Usage

write_xlsx(
  x,
  path = tempfile(fileext = ".xlsx"),
  col_names = TRUE,
  format_headers = TRUE,
  use_zip64 = FALSE
)

Arguments

x
data frame or named list of data frames that will be sheets in the xlsx

path
a file name to write to

col_names
write column names at the top of the file?

format_headers
make the col_names in the xlsx centered and bold

use_zip64
use zip64 to enable support for 4GB+ xlsx files. Not all platforms can read this.

Details

Currently supports strings, numbers, booleans and dates. Formatting options may be added in future versions.

Examples

# Roundtrip example with single excel sheet named 'mysheet'
tmp <- write_xlsx(list(mysheet = iris))
readxl::read_xlsx(tmp)

[Package writexl version 1.3 Index]